Clean up structure, remove unused code, and improve docs
