--- src/collision/narrow_phase.rs
@@ -1,1075 +1,0 @@
-//! Computes contacts between entities.
-//!
-//! See [`NarrowPhasePlugin`].
-
-use std::marker::PhantomData;
-
-use crate::{
-    data_structures::graph::EdgeIndex,
-    dynamics::solver::{
-        contact::ContactConstraint, ContactConstraints, ContactSoftnessCoefficients,
-    },
-    prelude::*,
-};
-use bevy::{
-    ecs::{
-        intern::Interned,
-        schedule::{ExecutorKind, LogLevel, ScheduleBuildSettings, ScheduleLabel},
-        system::{StaticSystemParam, SystemParam, SystemParamItem},
-    },
-    prelude::*,
-};
-use bit_vec::BitVec;
-use broad_phase::BroadCollisionPairs;
-use dynamics::solver::SolverDiagnostics;
-use rayon::iter::{IndexedParallelIterator, IntoParallelRefMutIterator, ParallelIterator};
-
-/// Computes contacts between entities and generates contact constraints for them.
-///
-/// Collisions are only checked between entities contained in [`BroadCollisionPairs`],
-/// which is handled by the [`BroadPhasePlugin`].
-///
-/// The results of the narrow phase are added into [`Collisions`].
-/// By default, a [`ContactConstraint`] is also generated for each contact manifold
-/// and added to the [`ContactConstraints`] resource.
-///
-/// The plugin takes a collider type. This should be [`Collider`] for
-/// the vast majority of applications, but for custom collisión backends
-/// you may use any collider that implements the [`AnyCollider`] trait.
-pub struct NarrowPhasePlugin<C: AnyCollider, H: CollisionHooks = ()> {
-    schedule: Interned<dyn ScheduleLabel>,
-    /// If `true`, the narrow phase will generate [`ContactConstraint`]s
-    /// and add them to the [`ContactConstraints`] resource.
-    ///
-    /// Contact constraints are used by the [`SolverPlugin`] for solving contacts.
-    generate_constraints: bool,
-    _phantom: PhantomData<(C, H)>,
-}
-
-impl<C: AnyCollider, H: CollisionHooks> NarrowPhasePlugin<C, H> {
-    /// Creates a [`NarrowPhasePlugin`] with the schedule used for running its systems
-    /// and whether it should generate [`ContactConstraint`]s for the [`ContactConstraints`] resource.
-    ///
-    /// Contact constraints are used by the [`SolverPlugin`] for solving contacts.
-    ///
-    /// The default schedule is [`PhysicsSchedule`].
-    pub fn new(schedule: impl ScheduleLabel, generate_constraints: bool) -> Self {
-        Self {
-            schedule: schedule.intern(),
-            generate_constraints,
-            _phantom: PhantomData,
-        }
-    }
-}
-
-impl<C: AnyCollider, H: CollisionHooks> Default for NarrowPhasePlugin<C, H> {
-    fn default() -> Self {
-        Self::new(PhysicsSchedule, true)
-    }
-}
-
-impl<C: AnyCollider, H: CollisionHooks + 'static> Plugin for NarrowPhasePlugin<C, H>
-where
-    for<'w, 's> SystemParamItem<'w, 's, H>: CollisionHooks,
-{
-    fn build(&self, app: &mut App) {
-        // For some systems, we only want one instance, even if there are multiple
-        // NarrowPhasePlugin instances with different collider types.
-        let is_first_instance = !app.world().is_resource_added::<NarrowPhaseInitialized>();
-
-        app.init_resource::<NarrowPhaseInitialized>()
-            .init_resource::<NarrowPhaseConfig>()
-            .init_resource::<Collisions>()
-            .init_resource::<DefaultFriction>()
-            .init_resource::<DefaultRestitution>()
-            .register_type::<(NarrowPhaseConfig, DefaultFriction, DefaultRestitution)>();
-
-        app.add_event::<CollisionStarted>()
-            .add_event::<CollisionEnded>();
-
-        if self.generate_constraints {
-            app.init_resource::<ContactConstraints>();
-        }
-
-        app.configure_sets(
-            self.schedule,
-            (
-                NarrowPhaseSet::First,
-                NarrowPhaseSet::CollectCollisions,
-                NarrowPhaseSet::PostProcess,
-                NarrowPhaseSet::GenerateConstraints,
-                NarrowPhaseSet::Last,
-            )
-                .chain()
-                .in_set(PhysicsStepSet::NarrowPhase),
-        );
-
-        // Set up the PostProcessCollisions schedule for user-defined systems
-        // that filter and modify collisions.
-        app.edit_schedule(PostProcessCollisions, |schedule| {
-            schedule
-                .set_executor_kind(ExecutorKind::SingleThreaded)
-                .set_build_settings(ScheduleBuildSettings {
-                    ambiguity_detection: LogLevel::Error,
-                    ..default()
-                });
-        });
-
-        // Collect contacts into `Collisions`.
-        app.add_systems(
-            self.schedule,
-            collect_collisions::<C, H>
-                .in_set(NarrowPhaseSet::CollectCollisions)
-                // Allowing ambiguities is required so that it's possible
-                // to have multiple collision backends at the same time.
-                .ambiguous_with_all(),
-        );
-
-        if self.generate_constraints {
-            // Generate contact constraints.
-            app.add_systems(
-                self.schedule,
-                generate_constraints::<C>
-                    .in_set(NarrowPhaseSet::GenerateConstraints)
-                    // Allowing ambiguities is required so that it's possible
-                    // to have multiple collision backends at the same time.
-                    .ambiguous_with_all(),
-            );
-        }
-
-        if is_first_instance {
-            app.add_systems(
-                self.schedule,
-                run_post_process_collisions_schedule.in_set(NarrowPhaseSet::PostProcess),
-            );
-        }
-    }
-
-    fn finish(&self, app: &mut App) {
-        // Register timer and counter diagnostics for collision detection.
-        app.register_physics_diagnostics::<CollisionDiagnostics>();
-    }
-}
-
-#[derive(Resource, Default)]
-struct NarrowPhaseInitialized;
-
-/// A resource for configuring the [narrow phase](NarrowPhasePlugin).
-#[derive(Resource, Reflect, Clone, Debug, PartialEq)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-#[cfg_attr(feature = "serialize", reflect(Serialize, Deserialize))]
-#[reflect(Debug, Resource, PartialEq)]
-pub struct NarrowPhaseConfig {
-    /// The default maximum [speculative margin](SpeculativeMargin) used for
-    /// [speculative collisions](dynamics::ccd#speculative-collision). This can be overridden
-    /// for individual entities with the [`SpeculativeMargin`] component.
-    ///
-    /// By default, the maximum speculative margin is unbounded, so contacts can be predicted
-    /// from any distance, provided that the bodies are moving fast enough. As the prediction distance
-    /// grows, the contact data becomes more and more approximate, and in rare cases, it can even cause
-    /// [issues](dynamics::ccd#caveats-of-speculative-collision) such as ghost collisions.
-    ///
-    /// By limiting the maximum speculative margin, these issues can be mitigated, at the cost
-    /// of an increased risk of tunneling. Setting it to `0.0` disables speculative collision
-    /// altogether for entities without [`SpeculativeMargin`].
-    ///
-    /// This is implicitly scaled by the [`PhysicsLengthUnit`].
-    ///
-    /// Default: `MAX` (unbounded)
-    pub default_speculative_margin: Scalar,
-
-    /// A contact tolerance that acts as a minimum bound for the [speculative margin](dynamics::ccd#speculative-collision).
-    ///
-    /// A small, positive contact tolerance helps ensure that contacts are not missed
-    /// due to numerical issues or solver jitter for objects that are in continuous
-    /// contact, such as pushing against each other.
-    ///
-    /// Making the contact tolerance too large will have a negative impact on performance,
-    /// as contacts will be computed even for objects that are not in close proximity.
-    ///
-    /// This is implicitly scaled by the [`PhysicsLengthUnit`].
-    ///
-    /// Default: `0.005`
-    pub contact_tolerance: Scalar,
-
-    /// If `true`, the current contacts will be matched with the previous contacts
-    /// based on feature IDs or contact positions, and the contact impulses from
-    /// the previous frame will be copied over for the new contacts.
-    ///
-    /// Using these impulses as the initial guess is referred to as *warm starting*,
-    /// and it can help the contact solver resolve overlap and stabilize much faster.
-    ///
-    /// Default: `true`
-    pub match_contacts: bool,
-}
-
-impl Default for NarrowPhaseConfig {
-    fn default() -> Self {
-        Self {
-            default_speculative_margin: Scalar::MAX,
-            contact_tolerance: 0.005,
-            match_contacts: true,
-        }
-    }
-}
-
-/// System sets for systems running in [`PhysicsStepSet::NarrowPhase`].
-#[derive(SystemSet, Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub enum NarrowPhaseSet {
-    /// Runs at the start of the narrow phase. Empty by default.
-    First,
-    /// Computes contacts between entities and adds them to the [`Collisions`] resource.
-    CollectCollisions,
-    /// Responsible for running the [`PostProcessCollisions`] schedule to allow user-defined systems
-    /// to filter and modify collisions.
-    ///
-    /// If you want to modify or remove collisions after [`NarrowPhaseSet::CollectCollisions`], you can
-    /// add custom systems to this set, or to [`PostProcessCollisions`].
-    PostProcess,
-    /// Generates [`ContactConstraint`]s and adds them to [`ContactConstraints`].
-    GenerateConstraints,
-    /// Runs at the end of the narrow phase. Empty by default.
-    Last,
-}
-
-fn collect_collisions<C: AnyCollider, H: CollisionHooks + 'static>(
-    mut narrow_phase: NarrowPhase<C>,
-    broad_collision_pairs: Res<BroadCollisionPairs>,
-    mut collision_started_event_writer: EventWriter<CollisionStarted>,
-    mut collision_ended_event_writer: EventWriter<CollisionEnded>,
-    time: Res<Time>,
-    hooks: StaticSystemParam<H>,
-    commands: ParallelCommands,
-    mut diagnostics: ResMut<CollisionDiagnostics>,
-) where
-    for<'w, 's> SystemParamItem<'w, 's, H>: CollisionHooks,
-{
-    let start = bevy::utils::Instant::now();
-
-    narrow_phase.update::<H>(
-        &broad_collision_pairs,
-        &mut collision_started_event_writer,
-        &mut collision_ended_event_writer,
-        time.delta_seconds_adjusted(),
-        &mut hooks.into_inner(),
-        commands,
-    );
-
-    diagnostics.narrow_phase = start.elapsed();
-    diagnostics.contact_count = narrow_phase.collisions.graph.edge_count() as u32;
-}
-
-// TODO: It'd be nice to generate the constraint in the same parallel loop as `collect_collisions`
-//       to avoid the extra iteration and queries. This is possible, but it wouldn't work with the current
-//       `PostProcessCollisions` setup.
-fn generate_constraints<C: AnyCollider>(
-    narrow_phase: NarrowPhase<C>,
-    mut constraints: ResMut<ContactConstraints>,
-    contact_softness: Res<ContactSoftnessCoefficients>,
-    default_friction: Res<DefaultFriction>,
-    default_restitution: Res<DefaultRestitution>,
-    time: Res<Time>,
-    mut collision_diagnostics: ResMut<CollisionDiagnostics>,
-    solver_diagnostics: Option<ResMut<SolverDiagnostics>>,
-) {
-    let start = bevy::utils::Instant::now();
-
-    let delta_secs = time.delta_seconds_adjusted();
-
-    constraints.clear();
-
-    // TODO: Parallelize.
-    for contacts in narrow_phase.collisions.iter() {
-        let Ok([collider1, collider2]) = narrow_phase
-            .collider_query
-            .get_many([contacts.entity1, contacts.entity2])
-        else {
-            continue;
-        };
-
-        let body1_bundle = collider1
-            .parent
-            .and_then(|p| narrow_phase.body_query.get(p.get()).ok());
-        let body2_bundle = collider2
-            .parent
-            .and_then(|p| narrow_phase.body_query.get(p.get()).ok());
-        if let (Some((body1, rb_collision_margin1)), Some((body2, rb_collision_margin2))) = (
-            body1_bundle.map(|(body, rb_collision_margin1, _)| (body, rb_collision_margin1)),
-            body2_bundle.map(|(body, rb_collision_margin2, _)| (body, rb_collision_margin2)),
-        ) {
-            // At least one of the bodies must be dynamic for contact constraints
-            // to be generated.
-            if !body1.rb.is_dynamic() && !body2.rb.is_dynamic() {
-                continue;
-            }
-
-            // Use the collider's own collision margin if specified, and fall back to the body's
-            // collision margin.
-            //
-            // The collision margin adds artificial thickness to colliders for performance
-            // and stability. See the `CollisionMargin` documentation for more details.
-            let collision_margin1 = collider1
-                .collision_margin
-                .or(rb_collision_margin1)
-                .map_or(0.0, |margin| margin.0);
-            let collision_margin2 = collider2
-                .collision_margin
-                .or(rb_collision_margin2)
-                .map_or(0.0, |margin| margin.0);
-            let collision_margin_sum = collision_margin1 + collision_margin2;
-
-            // Get combined friction and restitution coefficients of the colliders
-            // or the bodies they are attached to. Fall back to the global defaults.
-            let friction = collider1
-                .friction
-                .or(body1.friction)
-                .copied()
-                .unwrap_or(default_friction.0)
-                .combine(
-                    collider2
-                        .friction
-                        .or(body2.friction)
-                        .copied()
-                        .unwrap_or(default_friction.0),
-                );
-            let restitution = collider1
-                .restitution
-                .or(body1.restitution)
-                .copied()
-                .unwrap_or(default_restitution.0)
-                .combine(
-                    collider2
-                        .restitution
-                        .or(body2.restitution)
-                        .copied()
-                        .unwrap_or(default_restitution.0),
-                );
-
-            // Generate contact constraints for the computed contacts
-            // and add them to `constraints`.
-            narrow_phase.generate_constraints(
-                contacts,
-                &mut constraints,
-                &body1,
-                &body2,
-                &collider1,
-                &collider2,
-                friction,
-                restitution,
-                collision_margin_sum,
-                *contact_softness,
-                delta_secs,
-            );
-        }
-    }
-
-    collision_diagnostics.generate_constraints = start.elapsed();
-
-    if let Some(mut solver_diagnostics) = solver_diagnostics {
-        solver_diagnostics.contact_constraint_count = constraints.len() as u32;
-    }
-}
-
-/// A system parameter for managing the narrow phase.
-///
-/// The narrow phase computes contacts for each intersection pair
-/// in [`BroadCollisionPairs`], adds them to the [`Collisions`] resource,
-/// and generates [`ContactConstraints`] for the contacts.
-#[derive(SystemParam)]
-pub struct NarrowPhase<'w, 's, C: AnyCollider> {
-    parallel_commands: ParallelCommands<'w, 's>,
-    collider_query: Query<'w, 's, ColliderQuery<C>, Without<ColliderDisabled>>,
-    colliding_entities_query: Query<'w, 's, &'static mut CollidingEntities>,
-    body_query: Query<
-        'w,
-        's,
-        (
-            RigidBodyQueryReadOnly,
-            Option<&'static CollisionMargin>,
-            Option<&'static SpeculativeMargin>,
-        ),
-        Without<RigidBodyDisabled>,
-    >,
-    /// Contacts found by the narrow phase.
-    pub collisions: ResMut<'w, Collisions>,
-    /// Configuration options for the narrow phase.
-    pub config: Res<'w, NarrowPhaseConfig>,
-    length_unit: Res<'w, PhysicsLengthUnit>,
-    // These are scaled by the length unit.
-    default_speculative_margin: Local<'s, Scalar>,
-    contact_tolerance: Local<'s, Scalar>,
-}
-
-impl<C: AnyCollider> NarrowPhase<'_, '_, C> {
-    /// Updates the narrow phase by computing [`Contacts`] based on [`BroadCollisionPairs`]
-    /// and adding them to [`Collisions`].
-    fn update<H: CollisionHooks>(
-        &mut self,
-        broad_collision_pairs: &BroadCollisionPairs,
-        collision_started_event_writer: &mut EventWriter<CollisionStarted>,
-        collision_ended_event_writer: &mut EventWriter<CollisionEnded>,
-        delta_secs: Scalar,
-        hooks: &mut H::Item<'_, '_>,
-        commands: ParallelCommands,
-    ) where
-        for<'w, 's> SystemParamItem<'w, 's, H>: CollisionHooks,
-    {
-        // TODO: These scaled versions could be in their own resource
-        //       and updated just before physics every frame.
-        // Cache default margins scaled by the length unit.
-        if self.config.is_changed() {
-            *self.default_speculative_margin =
-                self.length_unit.0 * self.config.default_speculative_margin;
-            *self.contact_tolerance = self.length_unit.0 * self.config.contact_tolerance;
-        }
-
-        // Add new contact pairs to the contact graph.
-        for (entity1, entity2) in broad_collision_pairs.iter() {
-            self.collisions
-                .insert_collision_pair(Contacts::new(*entity1, *entity2));
-        }
-
-        // Compute contacts for all contact pairs.
-        let bit_vec = self.update_contacts::<H>(delta_secs, hooks, commands);
-
-        // Contact pairs that should be removed.
-        // This temporary storage is currently needed because removing pairs while iterating
-        // over the bit vec can invalidate the indices. If we figure out a stable mapping
-        // between contact pair indices and bits, we can remove this.
-        // TODO: Pre-allocate this with some reasonable capacity?
-        let mut to_remove = Vec::new();
-
-        // Process contact state changes, iterating over set bits.
-        // Set bits are found with the "count trailing zeros" method: https://lemire.me/blog/2018/02/21/iterating-over-set-bits-quickly/
-        for (i, mut bits) in bit_vec.blocks().enumerate() {
-            while bits != 0 {
-                let trailing_zeros = bits.trailing_zeros();
-                let contact_id = i * 32 + trailing_zeros as usize;
-
-                let contact_pair = self
-                    .collisions
-                    .graph
-                    .edge_weight_mut(EdgeIndex(contact_id as u32))
-                    .unwrap_or_else(|| {
-                        panic!("Contact pair not found for EdgeIndex({contact_id})")
-                    });
-
-                if contact_pair.flags.contains(ContactPairFlags::DISJOINT_AABB) {
-                    // Send a collision ended event if the contact pair was touching.
-                    let send_event = contact_pair
-                        .flags
-                        .contains(ContactPairFlags::TOUCHING | ContactPairFlags::CONTACT_EVENTS);
-                    if send_event {
-                        collision_ended_event_writer
-                            .send(CollisionEnded(contact_pair.entity1, contact_pair.entity2));
-                    }
-
-                    // Remove from `CollidingEntities`.
-                    Self::remove_colliding_entities(
-                        &mut self.colliding_entities_query,
-                        contact_pair.entity1,
-                        contact_pair.entity2,
-                    );
-
-                    // Remove the contact pair.
-                    let (entity1, entity2) = (contact_pair.entity1, contact_pair.entity2);
-                    to_remove.push((entity1, entity2));
-                } else if contact_pair
-                    .flags
-                    .contains(ContactPairFlags::STARTED_TOUCHING)
-                {
-                    // TODO: When sensor events are separate, handle them separately.
-
-                    // Send collision started event.
-                    if contact_pair
-                        .flags
-                        .contains(ContactPairFlags::CONTACT_EVENTS)
-                    {
-                        collision_started_event_writer
-                            .send(CollisionStarted(contact_pair.entity1, contact_pair.entity2));
-                    }
-
-                    // Add to `CollidingEntities`.
-                    Self::add_colliding_entities(
-                        &mut self.colliding_entities_query,
-                        contact_pair.entity1,
-                        contact_pair.entity2,
-                    );
-
-                    debug_assert!(!contact_pair.manifolds.is_empty());
-
-                    contact_pair
-                        .flags
-                        .set(ContactPairFlags::STARTED_TOUCHING, false);
-                } else if contact_pair
-                    .flags
-                    .contains(ContactPairFlags::STOPPED_TOUCHING)
-                {
-                    // Send collision ended event.
-                    if contact_pair
-                        .flags
-                        .contains(ContactPairFlags::CONTACT_EVENTS)
-                    {
-                        collision_ended_event_writer
-                            .send(CollisionEnded(contact_pair.entity1, contact_pair.entity2));
-                    }
-
-                    // Remove from `CollidingEntities`.
-                    Self::remove_colliding_entities(
-                        &mut self.colliding_entities_query,
-                        contact_pair.entity1,
-                        contact_pair.entity2,
-                    );
-
-                    debug_assert!(contact_pair.manifolds.is_empty());
-
-                    contact_pair
-                        .flags
-                        .set(ContactPairFlags::STOPPED_TOUCHING, false);
-                }
-
-                // Clear the least significant set bit.
-                bits &= bits - 1;
-            }
-        }
-
-        // Remove the contact pairs that were marked for removal.
-        for (entity1, entity2) in to_remove.drain(..) {
-            self.collisions.remove_collision_pair(entity1, entity2);
-        }
-    }
-
-    /// Adds the colliding entities to their respective `CollidingEntities` components.
-    fn add_colliding_entities(
-        query: &mut Query<&mut CollidingEntities>,
-        entity1: Entity,
-        entity2: Entity,
-    ) {
-        if let Ok(mut colliding_entities1) = query.get_mut(entity1) {
-            colliding_entities1.insert(entity2);
-        }
-        if let Ok(mut colliding_entities2) = query.get_mut(entity2) {
-            colliding_entities2.insert(entity1);
-        }
-    }
-
-    /// Removes the colliding entities from their respective `CollidingEntities` components.
-    fn remove_colliding_entities(
-        query: &mut Query<&mut CollidingEntities>,
-        entity1: Entity,
-        entity2: Entity,
-    ) {
-        if let Ok(mut colliding_entities1) = query.get_mut(entity1) {
-            colliding_entities1.remove(&entity2);
-        }
-        if let Ok(mut colliding_entities2) = query.get_mut(entity2) {
-            colliding_entities2.remove(&entity1);
-        }
-    }
-
-    fn update_contacts<H: CollisionHooks>(
-        &mut self,
-        delta_secs: Scalar,
-        hooks: &H::Item<'_, '_>,
-        par_commands: ParallelCommands,
-    ) -> BitVec
-    where
-        for<'w, 's> SystemParamItem<'w, 's, H>: CollisionHooks,
-    {
-        let initial = BitVec::from_elem(self.collisions.graph.edge_count(), false);
-        let state_change_bits = self
-            .collisions
-            .graph
-            .raw_edges_mut()
-            .par_iter_mut()
-            .enumerate()
-            .map_with(
-                initial.clone(),
-                |state_change_bits, (contact_id, contacts)| {
-                    let contacts = &mut contacts.weight;
-
-                    // TODO
-                    // self.handle_entity_pair(edge, &mut state_change_bits, edge_index, delta_secs);
-
-                    let Ok([collider1, collider2]) = self
-                        .collider_query
-                        .get_many([contacts.entity1, contacts.entity2])
-                    else {
-                        return state_change_bits.clone();
-                    };
-
-                    // Check if the AABBs of the colliders still overlap and the contact pair is valid.
-                    let overlap = collider1.aabb.intersects(&collider2.aabb);
-
-                    if !overlap {
-                        // The AABBs don't overlap. The contact pair will be removed later.
-                        contacts.flags.set(ContactPairFlags::DISJOINT_AABB, true);
-                        state_change_bits.set(contact_id, true);
-                    } else {
-                        // The AABBs overlap. Compute contacts.
-
-                        let body1_bundle = collider1
-                            .parent
-                            .and_then(|p| self.body_query.get(p.get()).ok());
-                        let body2_bundle = collider2
-                            .parent
-                            .and_then(|p| self.body_query.get(p.get()).ok());
-
-                        // The rigid body's collision margin and speculative margin will be used
-                        // if the collider doesn't have them specified.
-                        let (mut lin_vel1, rb_collision_margin1, rb_speculative_margin1) =
-                            body1_bundle
-                                .as_ref()
-                                .map(|(body, collision_margin, speculative_margin)| {
-                                    (
-                                        body.linear_velocity.0,
-                                        *collision_margin,
-                                        *speculative_margin,
-                                    )
-                                })
-                                .unwrap_or_default();
-                        let (mut lin_vel2, rb_collision_margin2, rb_speculative_margin2) =
-                            body2_bundle
-                                .as_ref()
-                                .map(|(body, collision_margin, speculative_margin)| {
-                                    (
-                                        body.linear_velocity.0,
-                                        *collision_margin,
-                                        *speculative_margin,
-                                    )
-                                })
-                                .unwrap_or_default();
-
-                        // Use the collider's own collision margin if specified, and fall back to the body's
-                        // collision margin.
-                        //
-                        // The collision margin adds artificial thickness to colliders for performance
-                        // and stability. See the `CollisionMargin` documentation for more details.
-                        let collision_margin1 = collider1
-                            .collision_margin
-                            .or(rb_collision_margin1)
-                            .map_or(0.0, |margin| margin.0);
-                        let collision_margin2 = collider2
-                            .collision_margin
-                            .or(rb_collision_margin2)
-                            .map_or(0.0, |margin| margin.0);
-                        let collision_margin_sum = collision_margin1 + collision_margin2;
-
-                        // Use the collider's own speculative margin if specified, and fall back to the body's
-                        // speculative margin.
-                        //
-                        // The speculative margin is used to predict contacts that might happen during the frame.
-                        // This is used for speculative collision. See the CCD and `SpeculativeMargin` documentation
-                        // for more details.
-                        let speculative_margin1 = collider1
-                            .speculative_margin
-                            .map_or(rb_speculative_margin1.map(|margin| margin.0), |margin| {
-                                Some(margin.0)
-                            });
-                        let speculative_margin2 = collider2
-                            .speculative_margin
-                            .map_or(rb_speculative_margin2.map(|margin| margin.0), |margin| {
-                                Some(margin.0)
-                            });
-
-                        // Compute the effective speculative margin, clamping it based on velocities and the maximum bound.
-                        let effective_speculative_margin = {
-                            let speculative_margin1 =
-                                speculative_margin1.unwrap_or(*self.default_speculative_margin);
-                            let speculative_margin2 =
-                                speculative_margin2.unwrap_or(*self.default_speculative_margin);
-                            let inv_delta_secs = delta_secs.recip();
-
-                            // Clamp velocities to the maximum speculative margins.
-                            if speculative_margin1 < Scalar::MAX {
-                                lin_vel1 =
-                                    lin_vel1.clamp_length_max(speculative_margin1 * inv_delta_secs);
-                            }
-                            if speculative_margin2 < Scalar::MAX {
-                                lin_vel2 =
-                                    lin_vel2.clamp_length_max(speculative_margin2 * inv_delta_secs);
-                            }
-
-                            // TODO: Check if AABBs intersect?
-
-                            // Compute the effective margin based on how much the bodies
-                            // are expected to move relative to each other.
-                            delta_secs * (lin_vel1 - lin_vel2).length()
-                        };
-
-                        // The maximum distance at which contacts are detected.
-                        // At least as large as the contact tolerance.
-                        let max_contact_distance = effective_speculative_margin
-                            .max(*self.contact_tolerance)
-                            + collision_margin_sum;
-
-                        let position1 = collider1.current_position();
-                        let position2 = collider2.current_position();
-
-                        // Initialize the contact pair.
-                        // TODO: Do this when creating contacts?
-                        contacts.body_entity1 = collider1.parent.map(|p| p.get());
-                        contacts.body_entity2 = collider2.parent.map(|p| p.get());
-                        contacts.flags.set(
-                            ContactPairFlags::SENSOR,
-                            collider1.is_sensor || collider2.is_sensor,
-                        );
-                        contacts.flags.set(
-                            ContactPairFlags::MODIFY_CONTACTS,
-                            collider1
-                                .active_hooks()
-                                .union(collider2.active_hooks())
-                                .contains(ActiveCollisionHooks::MODIFY_CONTACTS),
-                        );
-
-                        let was_touching = contacts.flags.contains(ContactPairFlags::TOUCHING);
-
-                        // Save the old manifolds for warm starting.
-                        let old_manifolds = contacts.manifolds.clone();
-
-                        // TODO: It'd be good to persist the manifolds and let Parry match contacts.
-                        //       This isn't currently done because it requires using Parry's contact manifold type.
-                        // Compute the contact manifolds using the effective speculative margin.
-                        contacts.manifolds = collider1.shape.contact_manifolds(
-                            collider2.shape,
-                            position1,
-                            *collider1.rotation,
-                            position2,
-                            *collider2.rotation,
-                            max_contact_distance,
-                        );
-
-                        // Check if the colliders are now touching.
-                        let mut touching = contacts.manifolds.iter().any(|m| !m.points.is_empty());
-
-                        if touching && contacts.flags.contains(ContactPairFlags::MODIFY_CONTACTS) {
-                            par_commands.command_scope(|mut commands| {
-                                touching = hooks.modify_contacts(contacts, &mut commands);
-                            });
-                            if !touching {
-                                contacts.manifolds.clear();
-                            }
-                        }
-
-                        contacts.flags.set(ContactPairFlags::TOUCHING, touching);
-
-                        // TODO: This condition is pretty arbitrary, mainly to skip dense trimeshes.
-                        //       If we let Parry handle contact matching, this wouldn't be needed.
-                        if contacts.manifolds.len() <= 4 && self.config.match_contacts {
-                            // TODO: Cache this?
-                            let distance_threshold = 0.1 * self.length_unit.0;
-
-                            for manifold in contacts.manifolds.iter_mut() {
-                                for previous_manifold in old_manifolds.iter() {
-                                    manifold.match_contacts(
-                                        &previous_manifold.points,
-                                        distance_threshold,
-                                    );
-                                }
-                            }
-                        }
-
-                        if touching && !was_touching {
-                            contacts.flags.set(ContactPairFlags::CONTACT_EVENTS, true);
-                            contacts.flags.set(ContactPairFlags::STARTED_TOUCHING, true);
-                            state_change_bits.set(contact_id, true);
-                        } else if !touching && was_touching {
-                            contacts.flags.set(ContactPairFlags::STOPPED_TOUCHING, true);
-                            state_change_bits.set(contact_id, true);
-                        }
-                    };
-
-                    state_change_bits.clone()
-                },
-            )
-            .reduce(
-                || initial.clone(),
-                |mut a, b| {
-                    a.or(&b);
-                    a
-                },
-            );
-
-        state_change_bits
-    }
-
-    /// Returns the [`Contacts`] between `entity1` and `entity2` if they are intersecting
-    /// or expected to start intersecting within the next frame. This includes
-    /// [speculative collision](dynamics::ccd#speculative-collision).
-    #[allow(clippy::too_many_arguments)]
-    pub fn handle_entity_pair(
-        &self,
-        contacts: &mut Contacts,
-        state_change_bits: &mut BitVec,
-        contact_id: usize,
-        delta_secs: Scalar,
-    ) {
-        let Ok([collider1, collider2]) = self
-            .collider_query
-            .get_many([contacts.entity1, contacts.entity2])
-        else {
-            return;
-        };
-
-        let body1_bundle = collider1
-            .parent
-            .and_then(|p| self.body_query.get(p.get()).ok());
-        let body2_bundle = collider2
-            .parent
-            .and_then(|p| self.body_query.get(p.get()).ok());
-
-        // The rigid body's collision margin and speculative margin will be used
-        // if the collider doesn't have them specified.
-        let (mut lin_vel1, rb_collision_margin1, rb_speculative_margin1) = body1_bundle
-            .as_ref()
-            .map(|(body, collision_margin, speculative_margin)| {
-                (
-                    body.linear_velocity.0,
-                    *collision_margin,
-                    *speculative_margin,
-                )
-            })
-            .unwrap_or_default();
-        let (mut lin_vel2, rb_collision_margin2, rb_speculative_margin2) = body2_bundle
-            .as_ref()
-            .map(|(body, collision_margin, speculative_margin)| {
-                (
-                    body.linear_velocity.0,
-                    *collision_margin,
-                    *speculative_margin,
-                )
-            })
-            .unwrap_or_default();
-
-        // Use the collider's own collision margin if specified, and fall back to the body's
-        // collision margin.
-        //
-        // The collision margin adds artificial thickness to colliders for performance
-        // and stability. See the `CollisionMargin` documentation for more details.
-        let collision_margin1 = collider1
-            .collision_margin
-            .or(rb_collision_margin1)
-            .map_or(0.0, |margin| margin.0);
-        let collision_margin2 = collider2
-            .collision_margin
-            .or(rb_collision_margin2)
-            .map_or(0.0, |margin| margin.0);
-        let collision_margin_sum = collision_margin1 + collision_margin2;
-
-        // Use the collider's own speculative margin if specified, and fall back to the body's
-        // speculative margin.
-        //
-        // The speculative margin is used to predict contacts that might happen during the frame.
-        // This is used for speculative collision. See the CCD and `SpeculativeMargin` documentation
-        // for more details.
-        let speculative_margin1 = collider1
-            .speculative_margin
-            .map_or(rb_speculative_margin1.map(|margin| margin.0), |margin| {
-                Some(margin.0)
-            });
-        let speculative_margin2 = collider2
-            .speculative_margin
-            .map_or(rb_speculative_margin2.map(|margin| margin.0), |margin| {
-                Some(margin.0)
-            });
-
-        // Compute the effective speculative margin, clamping it based on velocities and the maximum bound.
-        let effective_speculative_margin = {
-            let speculative_margin1 =
-                speculative_margin1.unwrap_or(*self.default_speculative_margin);
-            let speculative_margin2 =
-                speculative_margin2.unwrap_or(*self.default_speculative_margin);
-            let inv_delta_secs = delta_secs.recip();
-
-            // Clamp velocities to the maximum speculative margins.
-            if speculative_margin1 < Scalar::MAX {
-                lin_vel1 = lin_vel1.clamp_length_max(speculative_margin1 * inv_delta_secs);
-            }
-            if speculative_margin2 < Scalar::MAX {
-                lin_vel2 = lin_vel2.clamp_length_max(speculative_margin2 * inv_delta_secs);
-            }
-
-            // TODO: Check if AABBs intersect?
-
-            // Compute the effective margin based on how much the bodies
-            // are expected to move relative to each other.
-            delta_secs * (lin_vel1 - lin_vel2).length()
-        };
-
-        // The maximum distance at which contacts are detected.
-        // At least as large as the contact tolerance.
-        let max_contact_distance =
-            effective_speculative_margin.max(*self.contact_tolerance) + collision_margin_sum;
-
-        self.compute_contact_pair(
-            contacts,
-            state_change_bits,
-            contact_id,
-            &collider1,
-            &collider2,
-            max_contact_distance,
-        );
-    }
-
-    /// Computes contacts between `collider1` and `collider2`.
-    /// Returns `None` if no contacts are found.
-    ///
-    /// The given `max_distance` determines the maximum distance for a contact
-    /// to be detected. A value greater than zero means that contacts are generated
-    /// based on the closest points even if the shapes are separated.
-    #[allow(clippy::type_complexity, clippy::too_many_arguments)]
-    pub fn compute_contact_pair(
-        &self,
-        contacts: &mut Contacts,
-        state_change_bits: &mut BitVec,
-        contact_id: usize,
-        collider1: &ColliderQueryItem<C>,
-        collider2: &ColliderQueryItem<C>,
-        max_distance: Scalar,
-    ) {
-        let position1 = collider1.current_position();
-        let position2 = collider2.current_position();
-
-        // Check if the AABBs of the colliders still overlap and the contact pair is valid.
-        let overlap = collider1.aabb.intersects(&collider2.aabb);
-
-        if !overlap {
-            // The AABBs don't overlap. The contact pair will be removed later.
-            contacts.flags.set(ContactPairFlags::DISJOINT_AABB, true);
-            contacts.flags.set(ContactPairFlags::TOUCHING, false);
-            state_change_bits.set(contact_id, true);
-        } else {
-            // The AABBs overlap. Compute contacts.
-
-            let was_touching = contacts.flags.contains(ContactPairFlags::TOUCHING);
-
-            // TODO: It'd be good to persist the manifolds and let Parry match contacts.
-            //       This isn't currently done because it requires using Parry's contact manifold type.
-            // Compute the contact manifolds using the effective speculative margin.
-            let mut new_manifolds = collider1.shape.contact_manifolds(
-                collider2.shape,
-                position1,
-                *collider1.rotation,
-                position2,
-                *collider2.rotation,
-                max_distance,
-            );
-
-            // Check if the colliders are now touching.
-            let touching = !new_manifolds.is_empty();
-            contacts.flags.set(ContactPairFlags::TOUCHING, touching);
-
-            // TODO: This condition is pretty arbitrary, mainly to skip dense trimeshes.
-            //       If we let Parry handle contact matching, this wouldn't be needed.
-            if new_manifolds.len() <= 4 && self.config.match_contacts {
-                // TODO: Cache this?
-                let distance_threshold = 0.1 * self.length_unit.0;
-
-                for manifold in new_manifolds.iter_mut() {
-                    for previous_manifold in contacts.manifolds.iter() {
-                        manifold.match_contacts(&previous_manifold.points, distance_threshold);
-                    }
-                }
-            }
-
-            // Update the contacts.
-            contacts.manifolds = new_manifolds;
-
-            if touching && !was_touching {
-                contacts.flags.set(ContactPairFlags::CONTACT_EVENTS, true);
-                contacts.flags.set(ContactPairFlags::STARTED_TOUCHING, true);
-                state_change_bits.set(contact_id, true);
-            } else if !touching && was_touching {
-                contacts.flags.set(ContactPairFlags::STOPPED_TOUCHING, true);
-                state_change_bits.set(contact_id, true);
-            }
-        }
-    }
-
-    /// Generates [`ContactConstraint`]s for the given bodies and their corresponding colliders
-    /// based on the given `contacts`. The constraints are added to the `constraints` vector.
-    ///
-    /// The `collision_margin` can be used to add artificial thickness to the colliders,
-    /// which can improve performance and stability in some cases. See [`CollisionMargin`]
-    /// for more details.
-    ///
-    /// The `contact_softness` is used to tune the damping and stiffness of the contact constraints.
-    #[allow(clippy::too_many_arguments)]
-    pub fn generate_constraints(
-        &self,
-        contacts: &Contacts,
-        constraints: &mut Vec<ContactConstraint>,
-        body1: &RigidBodyQueryReadOnlyItem,
-        body2: &RigidBodyQueryReadOnlyItem,
-        collider1: &ColliderQueryItem<C>,
-        collider2: &ColliderQueryItem<C>,
-        friction: Friction,
-        restitution: Restitution,
-        collision_margin: impl Into<CollisionMargin> + Copy,
-        contact_softness: ContactSoftnessCoefficients,
-        delta_secs: Scalar,
-    ) {
-        let inactive1 = body1.rb.is_static() || body1.is_sleeping;
-        let inactive2 = body2.rb.is_static() || body2.is_sleeping;
-
-        // No collision response if both bodies are static or sleeping
-        // or if either of the colliders is a sensor collider.
-        if (inactive1 && inactive2)
-            || (collider1.is_sensor || body1.is_sensor)
-            || (collider2.is_sensor || body2.is_sensor)
-        {
-            return;
-        }
-
-        // When an active body collides with a sleeping body, wake up the sleeping body.
-        self.parallel_commands.command_scope(|mut commands| {
-            if body1.is_sleeping {
-                commands.queue(WakeUpBody(body1.entity));
-            } else if body2.is_sleeping {
-                commands.queue(WakeUpBody(body2.entity));
-            }
-        });
-
-        let contact_softness = if !body1.rb.is_dynamic() || !body2.rb.is_dynamic() {
-            contact_softness.non_dynamic
-        } else {
-            contact_softness.dynamic
-        };
-
-        // Generate contact constraints for each contact.
-        for (i, contact_manifold) in contacts.manifolds.iter().enumerate() {
-            let constraint = ContactConstraint::generate(
-                i,
-                contact_manifold,
-                body1,
-                body2,
-                collider1.entity,
-                collider2.entity,
-                collider1.transform.copied(),
-                collider2.transform.copied(),
-                collision_margin,
-                // TODO: Shouldn't this be the effective speculative margin?
-                *self.default_speculative_margin,
-                friction.dynamic_coefficient,
-                restitution.coefficient,
-                #[cfg(feature = "2d")]
-                contact_manifold.tangent_speed,
-                #[cfg(feature = "3d")]
-                contact_manifold.tangent_velocity,
-                contact_softness,
-                self.config.match_contacts,
-                delta_secs,
-            );
-
-            if !constraint.points.is_empty() {
-                constraints.push(constraint);
-            }
-        }
-    }
-}
-
-/// Runs the [`PostProcessCollisions`] schedule.
-fn run_post_process_collisions_schedule(world: &mut World) {
-    trace!("running PostProcessCollisions");
-    world.run_schedule(PostProcessCollisions);
-}
